Return TermQuery in case TermsQuery resolve down to one term
Summary: Generally TermsQuery will end up as ORed BooleanQuery of TermQueries. It would be efficient to result in a TermQuery if there is only one valid term after analyser chain is applied. This would also avoid the necessity to use TermQuery especially if the caller is not sure whether the analyser would break the terms further down, thus fixing the issue of 'TermQuery already set'.

Test Plan: Unit tested. Unit test files are on the way.

Reviewers: cpoersch, dcollins

Reviewed By: cpoersch

CC: mjustice, raiyenga

Differential Revision: https://all.phab.dev.bloomberg.com/D105430