bbsolr: Removing old Artifactory repo (no longer valid) and adding new repo
Summary:
The previous Artifactory repo (which we favoured) is no longer valid, and I have added in entries for the new Artifactory repo which we are using for maven builds.
I have left in the old Nexus repo since it still works, but we should really remove that too at some point. It isn't used in the resolver chain, but could be added again if necessary.

I added in the jcenter repo since that gave a correct version of org.restlet:2.1.1 which the sonatype repo had but seemed to be broken.
I left the sonatype and maven restlet repos out of the resolver chain since the Solr build worked without them, but they could be added if necessary.

Test Plan: With a clean ivy cache, I can "ant resolve" a complete lucene solr 4.8.1 build.

Reviewers: cpoersch, raiyenga, ksrivisa

Reviewed By: cpoersch

CC: mwashimi, mjustice

Differential Revision: https://all.phab.dev.bloomberg.com/D171172