Back port SOLR-6820.patch to our Solr.
Summary:
"The sync on the VersionInfo bucket in
DistributedUpdateProcesser#addDocument appears to be a large
bottleneck when using replication."

https://issues.apache.org/jira/browse/SOLR-6820

Test Plan:
Run on Alpha, raising the value from the default of 256 to 1000.
By processing the logs, compare before/after histograms for the time required for solr to
process updates. Main focus is on the long (~50sec) operations: do we see fewer of them? If there are fewer ~5sec operations that is also good.

Reviewers: raiyenga

Reviewed By: raiyenga

Differential Revision: https://all.phab.dev.bloomberg.com/D282610