bbsolr: Revert from sloppyFreq to freq for IntervalFilterQuery scoring
Summary: This an "alternative" fix to D244482, it leaves sloppyFreq() alone but prevents calling it (which was the problem sloppyFreq() iterated over the intervals, so advance()/nextDoc() were broken when they were called.

Test Plan: Tested using a single shard/core in dev and it seems to work.  No change in overall score compared with the D244482 fix.

Reviewers: raiyenga, ksrivisa, mjustice

Reviewed By: raiyenga

CC: cpoersch

Differential Revision: https://all.phab.dev.bloomberg.com/D367192