bbsolr: (NSRCH-553) Skip over null documents when writing response in TextResponseWriter:writeSolrDocumentList
Summary:
This just skips over any null documents that appear in the SolrDocumentList, instead of throwing an NPE which breaks the response.  Can't throw as we would then ignore the rest of the response (that's what the existing behaviour does).

This doesn't deal with the root cause which is why we are getting null docs at all.

Test Plan: Can rollout with next BBSolr train?  Could make it configurable using breg/property...

Reviewers: cpoersch, rhaddad15, dceccarelli4

Reviewed By: cpoersch

CC: raiyenga

Differential Revision: https://all.phab.dev.bloomberg.com/D476252