bbsolr-4.8.1-news: reinstate (lucene_solr_4_8_1 version of) solr/CHANGES.txt
Summary:
This reverts the solr/CHANGES.txt part of D261390 phabricator.

(Cherry-picking back from upstream usually causes conflicts with the older solr/CHANGES.txt and the */CHANGES.txt files are for the upstream releases really and so we generally haven't updated them when cherry-picking stuff back from upstream. Reverting this change to reduce files that need to be examined/ported to 5.5.x branch.)

Test Plan: -

Reviewers: vdalmia, dcollins, rhaddad15, raiyenga

Reviewed By: dcollins

Differential Revision: https://all.phab.dev.bloomberg.com/D537399